i18n: Move locale change method to the App
