+++ src/App.vue
@@ -1,13 +1,26 @@
 <template>
   <div id="app">
     <div id="nav">
-      <router-link to="/">
-        Home
-      </router-link>
+      <router-link to="/">Home</router-link>
     </div>
     <router-view />
   </div>
 </template>
+
+<script lang="ts">
+import Vue from 'vue';
+
+export default Vue.extend({
+  created(): void {
+    const locale = localStorage.getItem('locale');
+    if (locale) {
+      this.$i18n.locale = locale;
+    } else if (navigator.language) {
+      this.$i18n.locale = navigator.language.substring(0, 2);
+    }
+  },
+});
+</script>
 
 <style>
 #app {
