+++ src/components/I18Component.vue
@@ -2,7 +2,6 @@
   <div>
     <button @click="setLocale('en')">English</button>
     <button @click="setLocale('es')">Spanish</button>
-    <p>{{ $t('message.hello') }}</p>
     <p>{{ $d(new Date(), 'long') }}</p>
   </div>
 </template>
